Fix Int32 and Int16 read functions accidentally reading Int64
It was a copy-paste error.

The return value of the function should have been okay as I believe a
static cast chops off the left bits. However it is bad as we could have
been reading past the end of an allocated buffer.

Amends 9ec63a098db3a5493e763a31260ddeee8c67603f

+++ src/util/functions.h
@@ -84,7 +84,7 @@ template<typename Byte>
     requires(sizeof(Byte) == 1)
 inline Int32 ReadInt32(const Byte *buf, Uint32 off)
 {
-    return qFromBigEndian<Int64>(buf + off);
+    return qFromBigEndian<Int32>(buf + off);
 }
 
 template<typename Byte>
@@ -98,7 +98,7 @@ template<typename Byte>
     requires(sizeof(Byte) == 1)
 inline Int16 ReadInt16(const Byte *buf, Uint32 off)
 {
-    return qFromBigEndian<Int64>(buf + off);
+    return qFromBigEndian<Int16>(buf + off);
 }
 
 KTORRENT_EXPORT void UpdateCurrentTime();
